TST: Move most new loadtxt tests to its own file
This also adds two basic new tests around files/strings containing
the \0 character (prooving that we handle that gracefully).

Also adds tests for:
* the `_` thousands delimiter (should fail, but doesn't for float128
  right now)
* Failure modes when the number of rows changes (negative specifically)

Many of these tests came originally from Warren Weckesser and others
were added by Ross Barnowsky:
